Add Kodi module
This is NOT just the Kodi package. For that, just install regular kodi
as a package.

This module will set up the system so that Kodi launches on boot in GBM
mode, to allow for an optimized HTPC experience.

It is more or less equivalent to using LibreELEC, just that this is
NixOS.